Avoid clobbering ebx PIC register on i386
If you attempt to compile xv6 on an older compiler (e.g., default gcc
4.2 in OpenbSD), you will get the following error:

usertests.c:1553:3: error: PIC register clobbered by 'ebx' in 'asm'.

As I understand it, this is because we are compiling an i386 target,
and ebx is in use on such a target. Here
cf. https://gcc.gnu.org/bugzilla/show_bug.cgi?id=47602

The solution is to push and then pop the ebx around the assembly code
that uses it.

This allows xv6 to compile on OpenBSD; unfortunately, it still doesn't
boot in QEMU for me.